Move src/object to core and update tests
